Moving into Systems & Components Folders
