+++ Content.Client/Cargo/Systems/ClientPriceGunSystem.cs
@@ -1,6 +1,6 @@
 using Content.Shared.Cargo.Components;
-using Content.Shared.Timing;
 using Content.Shared.Cargo.Systems;
+using Content.Shared.Timing.Systems;
 
 namespace Content.Client.Cargo.Systems;
 
@@ -13,10 +13,7 @@ public sealed partial class ClientPriceGunSystem : SharedPriceGunSystem
 
     protected override bool GetPriceOrBounty(Entity<PriceGunComponent> entity, EntityUid target, EntityUid user)
     {
-        if (!TryComp(entity, out UseDelayComponent? useDelay) || _useDelay.IsDelayed((entity, useDelay)))
-            return false;
-
         // It feels worse if the cooldown is predicted but the popup isn't! So only do the cooldown reset on the server.
-        return true;
+        return _useDelay.IsDelayed(entity.Owner);
     }
 }
