+++ Content.Client/Implants/UI/ChameleonControllerBoundUserInterface.cs
@@ -1,8 +1,6 @@
-﻿using Content.Shared.Clothing;
-using Content.Shared.Implants;
-using Content.Shared.Preferences.Loadouts;
-using Content.Shared.Roles;
-using Content.Shared.Timing;
+﻿using Content.Shared.Implants;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 using JetBrains.Annotations;
 using Robust.Client.UserInterface;
 using Robust.Shared.Prototypes;
