+++ Content.Client/UserInterface/Systems/Atmos/GasTank/GasTankWindow.cs
@@ -4,7 +4,7 @@ using Content.Client.Resources;
 using Content.Client.Stylesheets;
 using Content.Shared.Atmos.Components;
 using Content.Shared.Atmos.EntitySystems;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Systems;
 using Robust.Client.Graphics;
 using Robust.Client.ResourceManagement;
 using Robust.Client.UserInterface;
