+++ Content.Client/UserInterface/Systems/Hands/HandsUIController.cs
@@ -6,7 +6,8 @@ using Content.Client.UserInterface.Systems.Hotbar.Widgets;
 using Content.Shared.Hands.Components;
 using Content.Shared.Input;
 using Content.Shared.Inventory.VirtualItem;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 using Robust.Client.Player;
 using Robust.Client.UserInterface;
 using Robust.Client.UserInterface.Controllers;
@@ -373,7 +374,8 @@ public sealed partial class HandsUIController : UIController, IOnStateEntered<Ga
                 hand.CooldownDisplay.Visible = false;
                 continue;
             }
-            var delay = _useDelay.GetLastEndingDelay((hand.Entity.Value, useDelay));
+
+            _useDelay.GetLastActiveDelay((hand.Entity.Value, useDelay), out var delay);
 
             hand.CooldownDisplay.Visible = true;
             hand.CooldownDisplay.FromTime(delay.StartTime, delay.EndTime);
