+++ Content.IntegrationTests/Tests/Storage/StorageInteractionTest.cs
@@ -4,7 +4,7 @@ using Content.IntegrationTests.Tests.Interaction;
 using Content.Shared.Input;
 using Content.Shared.PDA;
 using Content.Shared.Storage;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Components;
 using Robust.Client.UserInterface;
 using Robust.Shared.Containers;
 using Robust.Shared.GameObjects;
