+++ Content.Server/Cargo/Systems/PriceGunSystem.cs
@@ -2,8 +2,9 @@ using Content.Server.Popups;
 using Content.Server.Salvage.JobBoard;
 using Content.Shared.Cargo.Components;
 using Content.Shared.IdentityManagement;
-using Content.Shared.Timing;
 using Content.Shared.Cargo.Systems;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 using Robust.Shared.Audio.Systems;
 
 namespace Content.Server.Cargo.Systems;
