+++ Content.Server/DeviceLinking/Systems/PowerSensorSystem.cs
@@ -7,7 +7,8 @@ using Content.Shared.Interaction;
 using Content.Shared.NodeContainer;
 using Content.Shared.Popups;
 using Content.Shared.Power.Generator;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 using Content.Shared.Tools.Systems;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Map.Components;
@@ -67,8 +68,7 @@ public sealed partial class PowerSensorSystem : EntitySystem
             return;
 
         // no sound spamming
-        if (TryComp<UseDelayComponent>(uid, out var useDelay)
-            && !_useDelay.TryResetDelay((uid, useDelay), true))
+        if (!_useDelay.TryResetDelay(uid, true))
             return;
 
         // switch between input and output mode.
