+++ Content.Server/Fluids/EntitySystems/SpraySystem.cs
@@ -5,7 +5,6 @@ using Content.Shared.CCVar;
 using Content.Shared.Chemistry.EntitySystems;
 using Content.Shared.Fluids;
 using Content.Shared.Interaction;
-using Content.Shared.Timing;
 using Robust.Server.GameObjects;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Configuration;
@@ -13,6 +12,7 @@ using Robust.Shared.Physics.Components;
 using System.Numerics;
 using Content.Shared.Fluids.EntitySystems;
 using Content.Shared.Fluids.Components;
+using Content.Shared.Timing.Systems;
 using Robust.Server.Containers;
 using Robust.Shared.Audio;
 using Robust.Shared.Map;
