+++ Content.Server/Implants/ChameleonControllerSystem.cs
@@ -9,7 +9,7 @@ using Content.Shared.Preferences;
 using Content.Shared.Preferences.Loadouts;
 using Content.Shared.Roles;
 using Content.Shared.Station;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Systems;
 using Robust.Shared.Player;
 using Robust.Shared.Prototypes;
