+++ Content.Server/Singularity/EntitySystems/RadiationCollectorSystem.cs
@@ -1,7 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Content.Server.Atmos.EntitySystems;
-using Content.Server.Atmos.Components;
 using Content.Server.Popups;
 using Content.Server.Power.Components;
 using Content.Server.Power.EntitySystems;
@@ -12,7 +11,7 @@ using Content.Shared.Examine;
 using Content.Shared.Interaction;
 using Content.Shared.Radiation.Events;
 using Content.Shared.Singularity.Components;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Systems;
 using Robust.Shared.Containers;
 using Robust.Shared.Timing;
 
@@ -71,7 +70,7 @@ public sealed partial class RadiationCollectorSystem : EntitySystem
         if (!args.Complex)
             return;
 
-        if (TryComp(uid, out UseDelayComponent? useDelay) && !_useDelay.TryResetDelay((uid, useDelay), true))
+        if (!_useDelay.TryResetDelay(uid, true))
             return;
 
         ToggleCollector(uid, args.User, component);
