+++ Content.Shared/Friends/Systems/PettableFriendSystem.cs
@@ -4,7 +4,8 @@ using Content.Shared.Interaction.Events;
 using Content.Shared.NPC.Components;
 using Content.Shared.NPC.Systems;
 using Content.Shared.Popups;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 
 namespace Content.Shared.Friends.Systems;
