+++ Content.Shared/RepulseAttract/RepulseAttractSystem.cs
@@ -1,6 +1,5 @@
 ﻿using Content.Shared.Physics;
 using Content.Shared.Throwing;
-using Content.Shared.Timing;
 using Content.Shared.Weapons.Melee.Events;
 using Content.Shared.Whitelist;
 using Content.Shared.Wieldable;
@@ -8,6 +7,7 @@ using Robust.Shared.Map;
 using Robust.Shared.Physics.Components;
 using System.Numerics;
 using Content.Shared.RepulseAttract.Events;
+using Content.Shared.Timing.Systems;
 using Content.Shared.Weapons.Melee;
 
 namespace Content.Shared.RepulseAttract;
