+++ Content.Shared/StationTeleporter/SharedStationTeleporterSystem.cs
@@ -7,7 +7,8 @@ using Content.Shared.Power;
 using Content.Shared.Power.EntitySystems;
 using Content.Shared.StationTeleporter.Components;
 using Content.Shared.Teleportation.Systems;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Containers;
 
@@ -103,7 +104,7 @@ public abstract partial class SharedStationTeleporterSystem : EntitySystem
         if (args.Handled)
             return;
 
-        if (TryComp<UseDelayComponent>(args.Used, out var useDelayComp) && _useDelay.IsDelayed((args.Used, useDelayComp)))
+        if (_useDelay.IsDelayed(args.Used))
             return;
 
         _useDelay.TryResetDelay(args.Used);
