+++ Content.Shared/Timing/Components/UseDelayComponent.cs
@@ -1,15 +1,14 @@
+using Content.Shared.Timing.Systems;
 using Robust.Shared.GameStates;
 using Robust.Shared.Serialization;
 
-namespace Content.Shared.Timing;
+namespace Content.Shared.Timing.Components;
 
 /// <summary>
 /// Timer that creates a cooldown each time an object is activated/used.
 /// Can support additional, separate cooldown timers on the object by passing a unique ID with the system methods.
 /// </summary>
-[RegisterComponent]
-[NetworkedComponent]
-[Access(typeof(UseDelaySystem))]
+[RegisterComponent, NetworkedComponent, Access(typeof(UseDelaySystem))]
 public sealed partial class UseDelayComponent : Component
 {
     [DataField]
@@ -21,7 +20,7 @@ public sealed partial class UseDelayComponent : Component
     /// <remarks>
     /// This is only used at MapInit and should not be expected
     /// to reflect the length of the default delay after that.
-    /// Use <see cref="UseDelaySystem.TryGetDelayInfo"/> instead.
+    /// Use <see cref="Systems.UseDelaySystem.TryGetDelayInfo"/> instead.
     /// </remarks>
     [DataField]
     public TimeSpan Delay = TimeSpan.FromSeconds(1);
