+++ Content.Shared/Weapons/Melee/UseDelayOnMeleeHitSystem.cs
@@ -1,5 +1,6 @@
 ﻿using Content.Shared.Throwing;
-using Content.Shared.Timing;
+using Content.Shared.Timing.Components;
+using Content.Shared.Timing.Systems;
 using Content.Shared.Weapons.Melee.Components;
 using Content.Shared.Weapons.Melee.Events;
 
@@ -31,9 +32,6 @@ public sealed partial class UseDelayOnMeleeHitSystem : EntitySystem
     {
         var uid = ent.Owner;
 
-        if (!TryComp<UseDelayComponent>(uid, out var useDelay))
-            return;
-
-        _delay.TryResetDelay((uid, useDelay), checkDelayed: true);
+        _delay.TryResetDelay(uid, checkDelayed: true);
     }
 }
