+++ Content.Shared/Weapons/Ranged/Systems/UseDelayOnShootSystem.cs
@@ -1,11 +1,10 @@
-using Content.Shared.Timing;
 using Content.Shared.Weapons.Ranged.Components;
 
 namespace Content.Shared.Weapons.Ranged.Systems;
 
 public sealed partial class UseDelayOnShootSystem : EntitySystem
 {
-    [Dependency] private UseDelaySystem _delay = default!;
+    [Dependency] private Timing.Systems.UseDelaySystem _delay = default!;
 
     public override void Initialize()
     {
@@ -15,7 +14,6 @@ public sealed partial class UseDelayOnShootSystem : EntitySystem
 
     private void OnUseShoot(Entity<UseDelayOnShootComponent> ent, ref GunShotEvent args)
     {
-        if (TryComp(ent, out UseDelayComponent? useDelay))
-            _delay.TryResetDelay((ent, useDelay));
+        _delay.TryResetDelay(ent);
     }
 }
